requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0

pydantic==2.10.2
pydantic-settings==2.6.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12

pytest==8.3.3
httpx==0.27.2
email-validator>=2.1.0
